pyproject.toml
[tool.poetry]
name = "segmentation"
version = "0.1.0"
description = ""
authors = ["Václav Volhejn <vaclav.volhejn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pycocotools = "^2.0.7"
pillow = "^10.1.0"
tqdm = "^4.66.1"
jupyter = "^1.0.0"
scikit-image = "^0.22.0"
google-cloud-storage = "^2.14.0"
faiss-cpu = "^1.7.4"
plotly = "^5.18.0"
pandas = "^2.1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
